Revert "LPS-115615 Move PortalExecutor"
This reverts commit a928b6bbac9ac18c7587041e8878bd80adaa5d20.
