+++ modules/apps/data-engine/data-engine-service/src/main/java/com/liferay/data/engine/internal/nativeobject/DataEngineNativeObjectObserver.java
@@ -14,7 +14,6 @@
 
 package com.liferay.data.engine.internal.nativeobject;
 
-import com.liferay.data.engine.internal.petra.executor.DataEngineNativeObjectPortalExecutor;
 import com.liferay.data.engine.nativeobject.DataEngineNativeObject;
 import com.liferay.data.engine.nativeobject.DataEngineNativeObjectField;
 import com.liferay.data.engine.rest.dto.v2_0.DataDefinition;
